# OpenAPI 3.0.json
openapi: 3.0.3
info:
  title: Multi-tenant Commerce Ingestion API
  version: 1.0.0
  description: >
    CRUD + search APIs for Organizations, Websites, Orders, Fulfillments, and Tracking.
    Supports date-time range filtering and pagination (page/size).
servers:
  - url: https://api.example.com
...
paths:
  /organizations:
    post:
      tags: [Organizations]
      summary: Create organization
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [Organizations]
      summary: List/search organizations (date range + pagination)
      operationId: listOrganizations
      parameters:
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OrgStatus"
        - name: name
          in: query
          description: Name contains (case-insensitive, implementation-specific)
          schema:
            type: string
      responses:
        "200":
          description: Paged organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedOrganization"
        "400":
          $ref: "#/components/responses/BadRequest"
  /organizations/search:
    get:
      tags: [Organizations]
      summary: Search organizations by external/reference id
      operationId: searchOrganizationsByRef
      parameters:
        - name: externalId
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
      responses:
        "200":
          description: Paged organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedOrganization"
        "400":
          $ref: "#/components/responses/BadRequest"
  /organizations/{id}:
    get:
      tags: [Organizations]
      summary: Get organization by id
      operationId: getOrganizationById
      parameters:
        - $ref: "#/components/parameters/orgIdPath"
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Organizations]
      summary: Update organization (full replace)
      operationId: updateOrganization
      parameters:
        - $ref: "#/components/parameters/orgIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Organizations]
      summary: Update organization (partial)
      operationId: patchOrganization
      parameters:
        - $ref: "#/components/parameters/orgIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationPatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Organizations]
      summary: Delete organization
      operationId: deleteOrganization
      parameters:
        - $ref: "#/components/parameters/orgIdPath"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
...
  /fulfillments/{fulfillmentId}/tracking:
    post:
      tags: [Tracking]
      summary: Create tracking for a fulfillment
      operationId: createTracking
      parameters:
        - $ref: "#/components/parameters/fulfillmentIdPathStandalone"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackingCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tracking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [Tracking]
      summary: List/search tracking for a fulfillment (date range + pagination)
      operationId: listTracking
      parameters:
        - $ref: "#/components/parameters/fulfillmentIdPathStandalone"
        - $ref: "#/components/parameters/from"
        - $ref: "#/components/parameters/to"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/sort"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TrackingStatus"
        - name: carrier
          in: query
          schema:
            type: string
        - name: trackingNumber
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paged tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedTracking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /fulfillments/{fulfillmentId}/tracking/search:
    get:
      tags: [Tracking]
      summary: Search tracking by trackingNumber (within fulfillment)
      operationId: searchTrackingByNumber
      parameters:
        - $ref: "#/components/parameters/fulfillmentIdPathStandalone"
        - name: trackingNumber
          in: query
          required: true
          schema:
            type: string
        - name: carrier
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
      responses:
        "200":
          description: Paged tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedTracking"
        "400":
          $ref: "#/components/responses/BadRequest"
  /fulfillments/{fulfillmentId}/tracking/{trackingId}:
    get:
      tags: [Tracking]
      summary: Get tracking by id
      operationId: getTrackingById
      parameters:
        - $ref: "#/components/parameters/fulfillmentIdPathStandalone"
        - $ref: "#/components/parameters/trackingIdPath"
      responses:
        "200":
          description: Tracking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tracking"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Tracking]
      summary: Update tracking (full replace)
      operationId: updateTracking
      parameters:
        - $ref: "#/components/parameters/fulfillmentIdPathStandalone"
        - $ref: "#/components/parameters/trackingIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackingUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tracking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Tracking]
      summary: Update tracking (partial)
      operationId: patchTracking
      parameters:
        - $ref: "#/components/parameters/fulfillmentIdPathStandalone"
        - $ref: "#/components/parameters/trackingIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackingPatchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tracking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Tracking]
      summary: Delete tracking
      operationId: deleteTracking
      parameters:
        - $ref: "#/components/parameters/fulfillmentIdPathStandalone"
        - $ref: "#/components/parameters/trackingIdPath"
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    from:
      name: from
      in: query
      required: false
      description: Start datetime (inclusive), ISO-8601
      schema:
        type: string
        format: date-time
    to:
      name: to
      in: query
      required: false
      description: End datetime (inclusive), ISO-8601
      schema:
        type: string
        format: date-time
    page:
      name: page
      in: query
      required: false
      description: 0-based page index
      schema:
        type: integer
        minimum: 0
        default: 0
    size:
      name: size
      in: query
      required: false
      description: Batch size / page size
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    sort:
      name: sort
      in: query
      required: false
      description: Sorting, e.g. createdAt,desc or updatedAt,asc
      schema:
        type: string
        default: updatedAt,desc
    orgIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    fulfillmentIdPathStandalone:
      name: fulfillmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    trackingIdPath:
      name: trackingId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
      required: [timestamp, status, error, message, path]

    # ----------------------------
    # Enums
    # ----------------------------
    OrgStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
...
    TrackingStatus:
      type: string
      enum: [LABEL_CREATED, IN_TRANSIT, OUT_FOR_DELIVERY, DELIVERED, EXCEPTION, UNKNOWN]

    # ----------------------------
    # Organization
    # ----------------------------
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
          nullable: true
        name:
          type: string
        status:
          $ref: "#/components/schemas/OrgStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, status, createdAt, updatedAt]
    OrganizationCreateRequest:
      type: object
      properties:
        externalId:
          type: string
          nullable: true
        name:
          type: string
          minLength: 2
        status:
          $ref: "#/components/schemas/OrgStatus"
      required: [name]
    OrganizationUpdateRequest:
      type: object
      properties:
        externalId:
          type: string
          nullable: true
        name:
          type: string
          minLength: 2
        status:
          $ref: "#/components/schemas/OrgStatus"
      required: [name, status]
    OrganizationPatchRequest:
      type: object
      properties:
        externalId:
          type: string
          nullable: true
        name:
          type: string
          minLength: 2
        status:
          $ref: "#/components/schemas/OrgStatus"
    PagedOrganization:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
      required: [data, page, size, totalElements, totalPages, hasNext]

    # ----------------------------
    # Website
    # ----------------------------
...
    Tracking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fulfillmentId:
          type: string
          format: uuid
        trackingNumber:
          type: string
        carrier:
          type: string
          nullable: true
        trackingUrl:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TrackingStatus"
        isPrimary:
          type: boolean
        lastEventAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, fulfillmentId, trackingNumber, status, isPrimary, createdAt, updatedAt]
    TrackingCreateRequest:
      type: object
      properties:
        trackingNumber:
          type: string
        carrier:
          type: string
          nullable: true
        trackingUrl:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TrackingStatus"
        isPrimary:
          type: boolean
          default: false
        lastEventAt:
          type: string
          format: date-time
          nullable: true
      required: [trackingNumber]
    TrackingUpdateRequest:
      allOf:
        - $ref: "#/components/schemas/TrackingCreateRequest"
    TrackingPatchRequest:
      type: object
      properties:
        carrier:
          type: string
          nullable: true
        trackingUrl:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TrackingStatus"
        isPrimary:
          type: boolean
        lastEventAt:
          type: string
          format: date-time
          nullable: true
    PagedTracking:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Tracking"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
      required: [data, page, size, totalElements, totalPages, hasNext]